requirements.txt
Flask==2.2.2
Flask_Cors==3.0.10
keras==2.10.0
numpy==1.23.4
Pillow==9.4.0
waitress==2.1.2
opencv-python==4.7.0.68
tensorflow==2.10.0
